Trigger inventory update when adding new inventory
