+++ src/screens/Inventory/AddInventory.tsx
@@ -8,10 +8,11 @@ import BaseScreen from '../../components/BaseComponents/BaseScreen';
 import Button from '../../components/Button';
 import TextField from '../../components/TextField';
 import { InventoryRecord } from '../../lib/airtable/interface';
-import { createInventory } from '../../lib/airtable/request';
+import { createInventory, createInventoryUpdateAndUpdateInventory } from '../../lib/airtable/request';
 import { setCurrentInventoryIdInRedux } from '../../lib/redux/inventoryData';
 import { EMPTY_INVENTORY, selectAllCurrentSiteInventoryArray, selectAllProducts } from '../../lib/redux/inventoryDataSlice';
 import { getCurrentSiteId } from '../../lib/redux/siteData';
+import { selectCurrentUserId } from '../../lib/redux/userData';
 
 const styles = (theme: Theme) =>
   createStyles({
@@ -31,6 +32,7 @@ function AddInventory (props: AddInventoryProps) {
   const { classes } = props;
   const products = useSelector(selectAllProducts);
   const siteInventory = useSelector(selectAllCurrentSiteInventoryArray);
+  const userId = useSelector(selectCurrentUserId);
   const history = useHistory();
 
   // Product IDs for items that the site already has inventory for
@@ -55,6 +57,9 @@ function AddInventory (props: AddInventoryProps) {
     
     // createInventory returns the inventory item with an id
     inventory = await createInventory(inventory);
+
+    // Once the inventory record is created, trigger an inventory update.
+    await createInventoryUpdateAndUpdateInventory(userId, inventory, inventory.currentQuantity);
 
     // TODO: @wangannie create inventory update
